LayoutTests/imported/w3c/web-platform-tests/css/css-counter-styles/counter-style-symbols-syntax-expected.txt


FAIL @counter-style 'symbols: "X"' is valid assert_equals: expected (undefined) undefined but got (string) "\"X\""
FAIL @counter-style 'symbols: "X" "X"' is valid assert_equals: expected (undefined) undefined but got (string) "\"X\" \"X\""
FAIL @counter-style 'symbols: ident "X"' is valid assert_equals: expected (undefined) undefined but got (string) "ident \"X\""
FAIL @counter-style 'symbols: ident "X" url("foo.jpg")' is valid assert_equals: expected (undefined) undefined but got (string) "ident \"X\" url(\"foo.jpg\")"
PASS @counter-style 'symbols: ' is invalid
PASS @counter-style 'symbols: initial "X" "X"' is invalid
PASS @counter-style 'symbols: inherit "X" "X"' is invalid
PASS @counter-style 'symbols: unset "X" "X"' is invalid
